refactor(jose): extract jose header into its own definition
